+++ backend/database/migrations/2026_01_20_173303_create_order_items_table.php
@@ -14,7 +14,7 @@ return new class extends Migration {
             $table->foreignId('item_id')->constrained()->cascadeOnDelete();
 
             $table->unsignedInteger('quantity');
-            $table->integer('price_at_purchase');
+            $table->decimal('price_at_purchase', 10, 2);
 
             $table->timestamps();
